requirements.txt
alembic==1.7.7
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.4.0
itsdangerous==2.1.1
Jinja2==3.0.3
jsonschema==4.4.0
Mako==1.2.0
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyrsistent==0.18.1
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.32
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.7.0
